[164]: Create a p8e.go file in types to hold the uniquely p8e functionality being added (rather than mingling it in with the other stuff that we want to keep).
